# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TGS Admin API Documentation",
        "description": "Documentation de l'API REST pour le système TGS Admin - Toulouse Game Show",
        "contact": {
            "email": "support@tgs.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur de développement"
        },
        {
            "url": "https://votre-domaine.com",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/v1/salons/{salon}/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Liste des catégories",
                "description": "Récupère toutes les catégories d'un salon",
                "operationId": "getSalonCategories",
                "parameters": [
                    {
                        "name": "salon",
                        "in": "path",
                        "description": "ID du salon",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégories récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Invités"
                                                    },
                                                    "salon_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/salons/{salon}/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Détail d'une catégorie",
                "description": "Récupère les informations d'une catégorie spécifique",
                "operationId": "getSalonCategory",
                "parameters": [
                    {
                        "name": "salon",
                        "in": "path",
                        "description": "ID du salon",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "salon_id": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            }
        },
        "/api/v1/salons/{salon}/categories/{category}/articles": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Articles d'une catégorie",
                "description": "Récupère tous les articles appartenant à une catégorie",
                "operationId": "getCategoryArticles",
                "parameters": [
                    {
                        "name": "salon",
                        "in": "path",
                        "description": "ID du salon",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles de la catégorie récupérés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            }
        },
...
    },
...
}